Added generating functionality. Tested(not all cases).

+++ userdata.go
@@ -4,7 +4,7 @@ type UserData struct{
 
 	ID int64 `json:"id"`
 	DayToRemind byte `json:"daytoremind"`
-	HourToRemind byte `json:"hourtoremind"`
+	HourToRemind int `json:"hourtoremind"`
 	Name string `json:"name"`
 	Group string `json:"group"`
 	StartPulse int `json:"startpulse"`
